backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Validation & Settings
pydantic==2.9.0
pydantic-settings==2.5.0

# Task Queue
celery[redis]==5.4.0
redis==5.1.0

# HTTP Client
httpx==0.27.0

# Scraping & Parsing
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
trafilatura==1.12.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.9.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
